requirements.txt
aiohttp==3.9.1
aiohttp-retry==2.8.3
beautifulsoup4==4.12.2
lxml==4.9.3
pandas==2.1.3
openpyxl==3.1.2
aiofiles==23.2.1
loguru==0.7.2
pydantic==2.5.2
python-dotenv==1.0.0
playwright==1.40.0
